Refactored public appliances into separate app
