+++ django-cloudlaunch/baselaunch/views.py
@@ -109,29 +109,3 @@ class DeploymentTaskViewSet(viewsets.ModelViewSet):
         user = self.request.user
         return models.ApplicationDeploymentTask.objects.filter(
             deployment=deployment, deployment__owner=user)
-
-
-### Public Services ###
-class LocationViewSet(viewsets.ModelViewSet):
-    """
-    List of all locations
-    """
-    queryset = models.Location.objects.all()
-    serializer_class = serializers.LocationSerializer
-
-
-class SponsorViewSet(viewsets.ModelViewSet):
-    """
-    List sponsors
-    """
-    queryset = models.Sponsor.objects.all()
-    serializer_class = serializers.SponsorSerializer
-
-
-class PublicServiceViewSet(viewsets.ModelViewSet):
-    """
-    List public services
-    """
-    queryset = models.PublicService.objects.all()
-    serializer_class = serializers.PublicServiceSerializer
-
