--- django-cloudlaunch/cmlaunch/urls.py
@@ -1,7 +1,0 @@
-from django.conf.urls import url
-
-from . import views
-
-urlpatterns = [
-    url(r'^$', views.index, name='index'),
-]
