--- django-cloudlaunch/cmlaunch/views.py
@@ -1,5 +1,0 @@
-from django.http import HttpResponse
-
-
-def index(request):
-    return HttpResponse("Hello, world - from the new Cloudlaunch.")
