+++ django-cloudlaunch/public_appliances/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PublicAppliancesConfig(AppConfig):
+    name = 'publicappliances'
