+++ django-cloudlaunch/public_appliances/tests.py
